% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{angle}
\alias{angle}
\title{Compute the angle of a line, segment or intersection of two lines}
\usage{
angle(x, y, ...)
}
\arguments{
\item{x}{an object (point, segment, line)}

\item{y}{an object (point, segment, line)}

\item{degrees}{return angle in degrees if TRUE}
}
\value{
an angle in radians
}
\description{
Compute the angle of a line, segment or intersection of two lines
}
\examples{
# Angle of a line between two points
p1 <- point(1, 1)
p2 <- point(4, 5)
angle(p1, p2)
# Angle of a line segment
s1 <- segment(p1, p2)
angle(s1)
# Angle of a line
l1 <- s1@line
angle(l1)
# Angle between two lines
l2 <- line(slope = 2, intercept = 0)
angle(l1, l2)

}
