% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{arc}
\alias{arc}
\title{circle class}
\usage{
arc(
  center = S7::class_missing,
  radius = S7::class_missing,
  start = S7::class_missing,
  end = S7::class_missing,
  theta = S7::class_missing,
  n = S7::class_missing,
  style = style(),
  ...
)
}
\arguments{
\item{center}{point at center of the circle}

\item{radius}{distance between center and edge circle}

\item{n}{number of points in circle}

\item{style}{a style object}

\item{xy}{2-column matrix of points}

\item{<[`dynamic-dots`][rlang::dyn-dots]>}{arguments passed to style object if style is empty}
}
\description{
circle class
}
\examples{
# specify center point and radius
p <- point(0,0)
circle(p, radius = 6)

# make a regular polygon
xy <- circle(p, n = 6)@xy
plot(xy, asp = 1, type = "n")
polygon(xy)

}
