% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrows.R
\name{arrow_segment}
\alias{arrow_segment}
\title{arrow_segment class}
\usage{
arrow_segment(
  p1 = class_missing,
  p2 = class_missing,
  style = class_missing,
  ...
)
}
\arguments{
\item{p1}{point 1}

\item{p2}{point 2}

\item{style}{a style list}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> arguments passed to style object if style is empty}
}
\description{
arrow_segment class
}
