% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrows.R
\name{arrow_segment_list}
\alias{arrow_segment_list}
\title{arrow_segment_list}
\usage{
arrow_segment_list(.data = class_missing)
}
\arguments{
\item{.data}{list of arrow_segments}

\item{p1}{point_list of first points (p1)}

\item{p2}{point_list of second points (p2)}

\item{line}{line_list of lines associated with segments}

\item{style}{style_list of lines associated with segments}
}
\description{
Make lists of arrow_segments
}
\examples{
s1 <- segment(point(0,0), point(1,1))
s2 <- segment(point(1,1), point(0,1))
segment_list(c(s1,s2))
}
