% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{circle}
\alias{circle}
\title{circle class}
\usage{
circle(center = class_missing, radius = class_missing, n = class_missing)
}
\arguments{
\item{center}{point at center of the circle}

\item{radius}{distance between center and edge circle}

\item{n}{number of points in circle}

\item{xy}{2-column matrix of points}
}
\description{
circle class
}
\examples{
# specify center point and radius
p <- point(0,0)
circle(p, radius = 6)

# make a regular polygon
xy <- circle(p, n = 6)@xy
plot(xy, asp = 1, type = "n")
polygon(xy)

}
