% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{equation}
\alias{equation}
\title{equation}
\usage{
equation(x, ...)
}
\arguments{
\item{x}{object}

\item{type}{equation type. Can be \code{y}, \code{general}, or \code{parametric}}
}
\description{
equation
}
