% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpathdiagramr-package.R
\docType{package}
\name{ggpathdiagramr-package}
\alias{ggpathdiagramr}
\alias{ggpathdiagramr-package}
\title{ggpathdiagramr: Makes path diagrams}
\description{
Makes path diagrams for path analysis and structural equation models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wjschne/ggpathdiagramr}
  \item \url{https://wjschne.github.io/ggpathdiagramr/}
  \item Report bugs at \url{https://github.com/wjschne/ggpathdiagramr/issues}
}

}
\author{
\strong{Maintainer}: W. Joel Schneider \email{w.joel.schneider@gmail.com} (\href{https://orcid.org/0000-0002-8393-5316}{ORCID})

}
\keyword{internal}
