% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{\%|-\%}
\alias{\%|-\%}
\title{Find point perpendicular to 2 points}
\usage{
e1 \%|-\% e2
}
\arguments{
\item{e1}{first point}

\item{e2}{second point}
}
\description{
Find point perpendicular to 2 points
}
