% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{label}
\alias{label}
\title{label class}
\usage{
label(
  p = S7::class_missing,
  label = S7::class_missing,
  style = S7::class_missing,
  ...
)
}
\arguments{
\item{p}{point}

\item{label}{text label}

\item{style}{a style list}

\item{...}{properties passed to style}
}
\description{
label class
}
