% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{line}
\alias{line}
\title{line class}
\usage{
line(
  slope = S7::class_missing,
  intercept = S7::class_missing,
  x_intercept = S7::class_missing,
  a = S7::class_missing,
  b = S7::class_missing,
  c = S7::class_missing
)
}
\description{
line class
}
