% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{make_node}
\alias{make_node}
\title{Rotate an object in 2 dimensions}
\usage{
make_node(x, y, ...)
}
\arguments{
\item{x}{object}

\item{theta}{angle}

\item{center}{length 2 vector  or point about which rotation occurs}
}
\description{
Rotate an object in 2 dimensions
}
