% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{make_point}
\alias{make_point}
\title{Get x and y values for making points}
\usage{
make_point(x, y, theta, r, ...)
}
\arguments{
\item{x}{object}

\item{y}{numeric}

\item{theta}{angle or numeric}

\item{r}{numeric}
}
\description{
Get x and y values for making points
}
\keyword{internal}
