% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R, R/labels.R
\name{point}
\alias{point}
\alias{polar}
\alias{point_list}
\alias{label_list}
\title{point class}
\usage{
point(
  x = class_missing,
  y = class_missing,
  theta = class_missing,
  r = class_missing,
  style = class_missing,
  ...
)

polar(theta = class_missing, r = class_missing, style = class_missing, ...)

point_list(.data = class_missing)

label_list(.data = class_missing)
}
\arguments{
\item{x}{Coordinate on the x-axis}

\item{y}{Coordinate on the y-axis}

\item{theta}{Angle of the vector from the origin to the point}

\item{r}{Radius = Distance from the origin to the point}

\item{style}{a style object}

\item{...}{properties passed to style}
}
\description{
point class

point_list class

point_list class
}
