% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{projection}
\alias{projection}
\title{Find projection of a point on a line}
\usage{
projection(x, y, ...)
}
\arguments{
\item{x}{point}

\item{y}{line}
}
\description{
Find projection of a point on a line
}
