% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{radial_just}
\alias{radial_just}
\title{Create a style object with hjust and vjust specified via radial coordinates}
\usage{
radial_just(angle, multiplier, ...)
}
\arguments{
\item{angle}{numberic (radians), angle, or point (which contains angle and distance (multiplier))}

\item{multiplier}{make hjust and vjust larger}
}
\description{
Create a style object with hjust and vjust specified via radial coordinates
}
