% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{segment}
\alias{segment}
\title{segment class}
\usage{
segment(p1 = class_missing, p2 = class_missing)
}
\arguments{
\item{p1}{point 1}

\item{p2}{point 2}

\item{line}{line on which the segment lies}

\item{xy}{a 2 by 2 matrix of the coordinates of p1 and p2}
}
\description{
segment class
}
