% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{segment_list}
\alias{segment_list}
\title{segment_list}
\usage{
segment_list(.data = class_missing)
}
\arguments{
\item{.data}{list of segments}

\item{p1}{point_list of first points (p1)}

\item{p2}{point_list of second points (p2)}

\item{line}{line_list of lines associated with segments}
}
\description{
Make lists of segments
}
\examples{
s1 <- segment(point(0,0), point(1,1))
s2 <- segment(point(1,1), point(0,1))
segment_list(c(s1,s2))
}
