% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{slope}
\alias{slope}
\title{Find slope of a line or object associated with a line}
\usage{
slope(x, ...)
}
\arguments{
\item{x}{point, segment, line, or angle}
}
\description{
Find slope of a line or object associated with a line
}
