% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{style}
\alias{style}
\title{style class}
\usage{
style(
  alpha = class_missing,
  arrow_head = class_missing,
  arrow_fins = class_missing,
  angle = class_missing,
  color = class_missing,
  family = class_missing,
  fill = class_missing,
  fontface = class_missing,
  hjust = class_missing,
  lineheight = class_missing,
  linejoin = class_missing,
  linewidth = class_missing,
  linetype = class_missing,
  shape = class_missing,
  size = class_missing,
  size.unit = class_missing,
  vjust = class_missing
)
}
\arguments{
\item{alpha}{numeric value for alpha transparency}

\item{arrow_head}{A 2-column matrix of polygon points}

\item{arrow_fins}{A 2-column matrix of polygon points}

\item{angle}{angle of text}

\item{color}{character string for color}

\item{family}{font family}

\item{fill}{character string for fill color}

\item{fontface}{Can be plain, bold, italic, or bold.italic}

\item{hjust}{horizontal justification}

\item{lineheight}{height of line of text}

\item{linejoin}{height of line of text}

\item{linewidth}{width of lines}

\item{linetype}{type of lines}

\item{shape}{type of shape}

\item{size}{numeric size}

\item{size.unit}{How the size aesthetic is interpreted: as points ("pt"), millimetres ("mm"), centimetres ("cm"), inches ("in"), or picas ("pc").}

\item{vjust}{vertical justification}
}
\description{
style class
}
