% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.R
\name{angle}
\alias{angle}
\alias{angle_list}
\alias{degree}
\alias{radian}
\alias{turn}
\alias{gradian}
\title{angle class}
\usage{
angle(
  radian = class_missing,
  degree = class_missing,
  turn = class_missing,
  gradian = class_missing
)

angle_list(.data = class_missing)

degree(degree)

radian(radian)

turn(turn)

gradian(gradian)
}
\arguments{
\item{radian}{radians}

\item{degree}{degrees}

\item{turn}{proportion of full turns of a circle (1 turn = 2 * pi radians)}

\item{gradian}{gradians, gons, or grads (right angle = 100 gradians)}

\item{.data}{vector of angles}
}
\description{
Creates an angle in the metric of radians, degrees, turns, and gradians.
}
\details{
Angles are never negative and always less than a full turn (360 degrees). Thus, \code{angle(degree = -10)} returns an angle of 350, and \code{angle(degree = 360)} returns an angle of 0.
}
\examples{
# Four Different ways to make a right angle
## A quarter turn
angle(.25)

## half pi radians
angle(radian = .5 * pi)

## 90 degrees
angle(degree = 90)

## 100 gradians
angle(gradian = 100)

# Convenience wrappers
degree(90)
radian(.5 * pi)
gradian(100)
turn(.25)

# Operations
angle(degree = 30) + angle(degree = 20)
angle(degree = 350) + angle(degree = 20)
angle(degree = 30) - angle(degree = 20)
angle(degree = 30) - angle(degree = 50)

2 * angle(degree = 30)
angle(degree = 30) / 3

angle(degree = 30) + pi # added or subtracted numbers are radians
radian(1) + 1 # added or subtracted numbers are radians
degree(30) + 180 # added or subtracted numbers are degrees
turn(.25) + .25 # added or subtracted numbers are turns
gradian(50) + 50 # added or subtracted numbers are turns

Trigonometric functions work as normal
x <- angle(degree = 180)
sin(x)
cos(x)
tan(x)
# Two ways to make an angle_list
angle(c(0,1))
angle_list(c(angle(0), angle(1)))
}
