% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{angle}
\alias{angle}
\alias{angle_list}
\title{angle class}
\usage{
angle(
  turn = S7::class_missing,
  radian = S7::class_missing,
  degree = S7::class_missing,
  gradian = S7:class_missing
)

angle_list(.data = class_missing)
}
\arguments{
\item{turn}{proportion of full turns of a circle (1 turn = 2 * pi radians)}

\item{radian}{radians}

\item{degree}{degrees}

\item{gradian}{gradians, gons, or grads (right angle = 100 gradians)}
}
\description{
angle class

angle list class
}
\examples{
# Four Different ways to make a right angle
## A quarter turn
angle(.25)

## half pi radians
angle(radian = .5 * pi)

## 90 degrees
angle(degree = 90)

## 100 gradians
angle(gradian = 100)

# Operations
angle(degree = 30) + angle(degree = 20)
angle(degree = 350) + angle(degree = 20)
angle(degree = 30) - angle(degree = 20)
angle(degree = 30) - angle(degree = 50)
2 * angle(degree = 30)
angle(degree = 30) / 3
}
