% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{arc}
\alias{arc}
\title{arc class}
\usage{
arc(
  center = class_missing,
  radius = class_missing,
  start = class_missing,
  end = class_missing,
  theta = class_missing,
  n = class_missing,
  style = class_missing,
  ...
)
}
\arguments{
\item{center}{point at center of the circle}

\item{radius}{distance between center and edge circle}

\item{start}{start angle}

\item{end}{end angle}

\item{theta}{interior angle (end - start)}

\item{n}{number of points in circle}

\item{style}{a style object}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> arguments passed to style object if style is empty}

\item{xy}{2-column matrix of points}
}
\description{
arc class
}
\examples{
# specify center point and radius
p <- point(0,0)
circle(p, radius = 6)

# make a regular polygon
xy <- circle(p, n = 6)@xy
plot(xy, asp = 1, type = "n")
polygon(xy)

}
