% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{ellipse}
\alias{ellipse}
\title{an ellipse}
\usage{
ellipse(
  center = S7::class_missing,
  a = S7::class_missing,
  b = S7::class_missing,
  m1 = S7::class_missing,
  m2 = S7::class_missing,
  theta = S7::class_missing,
  n = S7::class_missing
)
}
\arguments{
\item{center}{point at center of ellipse}

\item{a}{distance of semi-major axis}

\item{b}{distance of semi-minor axis}

\item{m1}{exponent of semi-major axis}

\item{m2}{exponent of semi-minor axis}

\item{theta}{ellipse rotation}

\item{n}{number of polygon points}

\item{xy}{polygon points}

\item{style}{a style object}

\item{<[`dynamic-dots`][rlang::dyn-dots]>}{arguments passed to style object if style is empty}
}
\description{
an ellipse
}
