% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{rectangle}
\alias{rectangle}
\title{rectangle class}
\usage{
rectangle(
  center = S7::class_missing,
  width = S7::class_missing,
  height = S7::class_missing,
  northeast = S7::class_missing,
  northwest = S7::class_missing,
  southwest = S7::class_missing,
  southeast = S7::class_missing,
  style = S7::class_missing,
  ...
)
}
\arguments{
\item{center}{point at center of the circle}

\item{width}{width}

\item{height}{height}

\item{style}{a style object}

\item{xy}{2-column matrix of points}

\item{<[`dynamic-dots`][rlang::dyn-dots]>}{arguments passed to style object if style is empty}
}
\description{
rectangle class
}
\examples{
# specify center point and radius
p <- point(0,0)
rectangle(p, width = 2, height = 2)
}
