% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.R
\name{segment}
\alias{segment}
\title{segment class}
\usage{
segment(
  p1 = S7::class_missing,
  p2 = S7::class_missing,
  style = S7::class_missing,
  ...
)
}
\arguments{
\item{p1}{point 1}

\item{p2}{point 2}

\item{style}{a style list}

\item{...}{properties passed to style}

\item{line}{line on which the segment lies}

\item{xy}{a 2 by 2 matrix of the coordinates of p1 and p2}
}
\description{
segment class
}
