% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{style}
\alias{style}
\title{style class}
\usage{
style(
  alpha = class_missing,
  angle = class_missing,
  color = class_missing,
  family = class_missing,
  fill = class_missing,
  hjust = class_missing,
  linewidth = class_missing,
  linetype = class_missing,
  pch = class_missing,
  shape = class_missing,
  size = class_missing,
  vjust = class_missing
)
}
\arguments{
\item{alpha}{numeric value for alpha transparency}

\item{angle}{angle of text}

\item{color}{character string for color}

\item{family}{font family}

\item{fill}{character string for fill color}

\item{hjust}{horizontal justification}

\item{linewidth}{width of lines}

\item{linetype}{type of lines}

\item{pch}{number for point character}

\item{shape}{type of shape}

\item{size}{numeric size}

\item{vjust}{vertical justification}
}
\description{
style class
}
