% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{style_base}
\alias{style_base}
\alias{style}
\alias{style_point}
\alias{style_line}
\alias{style_arrow}
\alias{style_polygon}
\alias{style_label}
\title{style class}
\usage{
style_base(alpha = class_missing, color = class_missing)

style(
  alpha = class_missing,
  angle = class_missing,
  arrow_head = class_missing,
  arrow_fins = class_missing,
  arrow_mid = class_missing,
  color = class_missing,
  family = class_missing,
  fill = class_missing,
  fontface = class_missing,
  hjust = class_missing,
  justify = class_missing,
  label.color = class_missing,
  label.margin = class_missing,
  label.padding = class_missing,
  label.r = class_missing,
  label.size = class_missing,
  length = class_missing,
  length_head = class_missing,
  length_fins = class_missing,
  length_mid = class_missing,
  lineend = class_missing,
  lineheight = class_missing,
  linejoin = class_missing,
  linewidth_fins = class_missing,
  linewidth_head = class_missing,
  linewidth = class_missing,
  linetype = class_missing,
  nudge_x = class_missing,
  nudge_y = class_missing,
  polar_just_angle = class_missing,
  polar_just_multiplier = class_missing,
  resect = class_missing,
  resect_fins = class_missing,
  resect_head = class_missing,
  shape = class_missing,
  size = class_missing,
  size.unit = class_missing,
  stroke = class_missing,
  stroke_color = class_missing,
  stroke_width = class_missing,
  text.color = class_missing,
  vjust = class_missing,
  ...
)

style_point(
  alpha = class_missing,
  color = class_missing,
  fill = class_missing,
  shape = class_missing,
  size = class_missing,
  stroke = class_missing,
  ...
)

style_line(
  alpha = class_missing,
  color = class_missing,
  lineend = class_missing,
  linejoin = class_missing,
  linewidth = class_missing,
  linetype = class_missing,
  ...
)

style_arrow(
  alpha = class_missing,
  arrow_fins = class_missing,
  arrow_head = class_missing,
  arrow_mid = class_missing,
  color = class_missing,
  justify = class_missing,
  length_fins = class_missing,
  length_head = class_missing,
  length_mid = class_missing,
  lineend = class_missing,
  linejoin = class_missing,
  linewidth = class_missing,
  linewidth_fins = class_missing,
  linewidth_head = class_missing,
  linetype = class_missing,
  resect = class_missing,
  resect_fins = class_missing,
  resect_head = class_missing,
  ...
)

style_polygon(
  alpha = class_missing,
  color = class_missing,
  lineend = class_missing,
  linejoin = class_missing,
  linewidth = class_missing,
  linetype = class_missing,
  fill = class_missing,
  rule = class_missing,
  ...
)

style_label(
  alpha = class_missing,
  color = class_missing,
  angle = class_missing,
  family = class_missing,
  fill = class_missing,
  fontface = class_missing,
  hjust = class_missing,
  label.color = class_missing,
  label.margin = class_missing,
  label.padding = class_missing,
  label.r = class_missing,
  label.size = class_missing,
  lineheight = class_missing,
  polar_just_angle = class_missing,
  polar_just_multiplier = class_missing,
  nudge_x = class_missing,
  nudge_y = class_missing,
  size = class_missing,
  text.color = class_missing,
  vjust = class_missing,
  ...
)
}
\arguments{
\item{alpha}{numeric value for alpha transparency}

\item{color}{character string for color}

\item{angle}{angle of text}

\item{arrow_head}{A 2-column matrix of polygon points}

\item{arrow_fins}{A 2-column matrix of polygon points}

\item{arrow_mid}{A 2-column matrix of polygon points}

\item{family}{font family}

\item{fill}{character string for fill color}

\item{fontface}{Can be plain, bold, italic, or bold.italic}

\item{hjust}{horizontal justification}

\item{lineheight}{height of line of text}

\item{linejoin}{height of line of text}

\item{linewidth}{width of lines}

\item{linetype}{type of lines}

\item{shape}{type of shape}

\item{size}{numeric size}

\item{size.unit}{How the size aesthetic is interpreted: as points ("pt"), millimeters ("mm"), centimeters ("cm"), inches ("in"), or picas ("pc").}

\item{vjust}{vertical justification}
}
\description{
style class
}
